replaced avatar component with image component
